source/requirements.txt
attrs==21.2.0
certifi==2021.10.8
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
django==3.2.16
djangorestframework==3.12.2
Fiona==1.8.20
geopandas==0.10.2
munch==2.5.0
numpy==1.21.4
pandas==1.3.4
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.10.0
Shapely==1.8.0
six==1.16.0
matplotlib==3.3.4
PyYAML==5.3.1
Rtree==0.9.7
pysqlite3==0.4.6
SQLAlchemy==1.4.31
psycopg2==2.8.6
platformdirs==2.4.1
paramiko==2.7.2
python-dotenv==0.20.0
ezdxf==0.18.1
typing-extensions==4.3.0
APScheduler==3.9.1
gunicorn==20.1.0
